requirements.txt
pycryptodome
pandas
pytest
